refactor: cy separate different commands
