+++ packages/cypress/src/integration/profile.spec.ts
@@ -4,8 +4,8 @@ import { MESSAGE_MAX_CHARACTERS } from '../../../../src/pages/User/constants'
 import { missing } from '../../../../src/pages/User/impact/labels'
 import { contact } from '../../../../src/pages/User/labels'
 import { MOCK_DATA } from '../data'
-import { UserMenuItem } from '../support/commands'
 import { setIsPreciousPlastic } from '../utils/TestUtils'
+import { UserMenuItem } from '../support/commandsUi'
 
 const { admin, subscriber } = MOCK_DATA.users
 const eventReader = MOCK_DATA.users.event_reader
