+++ packages/cypress/src/support/commands.ts
@@ -7,12 +7,6 @@ import { Auth, TestDB } from './db/firebase'
 
 import type { firebase } from './db/firebase'
 
-export enum UserMenuItem {
-  Profile = 'Profile',
-  Settings = 'Settings',
-  LogOut = 'Logout',
-}
-
 declare global {
   namespace Cypress {
     interface Chainable {
@@ -48,25 +42,6 @@ declare global {
 
       step(message: string)
 
-      toggleUserMenuOn(): Chainable<void>
-      toggleUserMenuOff(): Chainable<void>
-
-      /**
-       * Trigger form validation
-       */
-      screenClick(): Chainable<void>
-
-      clickMenuItem(menuItem: UserMenuItem): Chainable<void>
-
-      /**
-       * Selecting options from the react-select picker can be a bit fiddly
-       * so user helper method to locate select box, type input and pick tag
-       * (if exists) https://github.com/cypress-io/cypress/issues/549
-       * @param tagname This will be typed into the input box and selected from the dropdown list
-       * @param selector Specify the selector of the react-select element
-       **/
-      selectTag(tagName: string, selector?: string): Chainable<void>
-
       interceptAddressSearchFetch(addressResponse): Chainable<void>
 
       interceptAddressReverseFetch(addressResponse): Chainable<void>
@@ -80,85 +55,51 @@ declare global {
  * @remark - async code should be wrapped in a Cypress.promise block to allow the resolved promise to be
  * used in chained results
  */
-const attachCustomCommands = (Cypress: Cypress.Cypress) => {
-  const firestore = TestDB
-  /** Delete an indexeddb - resolving true on success and false if blocked (open connections) */
-  Cypress.Commands.add('deleteIDB', (name: string) => {
-    cy.wrap('Delete Firebase IDB: ' + name)
-      .then(() => {
-        return new Cypress.Promise<boolean>((resolve) => {
-          // Ensure DB exists - NOTE - only supported in chrome
-          // ;(indexedDB as any).databases().then((names: string[]) => {
-          //   if (names.includes(name)) {
-          deleteDB(name, {
-            // blocked implies active connection; for now just resolve false but in the
-            // future may want to find better resolution
-            blocked: () => resolve(false),
-          })
-            .then(() => resolve(true))
-            .catch(() => resolve(false))
+const firestore = TestDB
+/** Delete an indexeddb - resolving true on success and false if blocked (open connections) */
+Cypress.Commands.add('deleteIDB', (name: string) => {
+  cy.wrap('Delete Firebase IDB: ' + name)
+    .then(() => {
+      return new Cypress.Promise<boolean>((resolve) => {
+        // Ensure DB exists - NOTE - only supported in chrome
+        // ;(indexedDB as any).databases().then((names: string[]) => {
+        //   if (names.includes(name)) {
+        deleteDB(name, {
+          // blocked implies active connection; for now just resolve false but in the
+          // future may want to find better resolution
+          blocked: () => resolve(false),
         })
+          .then(() => resolve(true))
+          .catch(() => resolve(false))
       })
-      .then((deleted) => cy.log('deleted?', deleted))
-  })
-
-  Cypress.Commands.add('setSessionStorage', (key: string, value: string) => {
-    cy.wrap(`setSessionStorage - ${key}:${value}`).then(() => {
-      cy.window().its('sessionStorage').invoke('setItem', key, value)
-      cy.window()
-        .its('sessionStorage')
-        .invoke('getItem', key)
-        .should('eq', value)
     })
-  })
+    .then((deleted) => cy.log('deleted?', deleted))
+})
 
-  Cypress.Commands.add('clearServiceWorkers', () => {
-    cy.window().then((w) => {
-      cy.wrap('Clearing service workers').then(() => {
-        return new Cypress.Promise((resolve) => {
-          // if running production builds locally may also need to remove service workers between runs
-          if (w.navigator && navigator.serviceWorker) {
-            navigator.serviceWorker.getRegistrations().then((registrations) => {
-              registrations.forEach((registration) => {
-                registration.unregister()
-              })
-              resolve()
+Cypress.Commands.add('setSessionStorage', (key: string, value: string) => {
+  cy.wrap(`setSessionStorage - ${key}:${value}`).then(() => {
+    cy.window().its('sessionStorage').invoke('setItem', key, value)
+    cy.window()
+      .its('sessionStorage')
+      .invoke('getItem', key)
+      .should('eq', value)
+  })
+})
+
+Cypress.Commands.add('clearServiceWorkers', () => {
+  cy.window().then((w) => {
+    cy.wrap('Clearing service workers').then(() => {
+      return new Cypress.Promise((resolve) => {
+        // if running production builds locally may also need to remove service workers between runs
+        if (w.navigator && navigator.serviceWorker) {
+          navigator.serviceWorker.getRegistrations().then((registrations) => {
+            registrations.forEach((registration) => {
+              registration.unregister()
             })
-          } else {
             resolve()
-          }
-        })
-      })
-    })
-  })
-
-  /**
-   * Login and logout commands use the sytem interface to log a user in or out
-   */
-  Cypress.Commands.add(
-    'login',
-    (email: string, password: string, checkUI = true) => {
-      Cypress.log({
-        displayName: 'login',
-        consoleProps: () => {
-          return { email, password }
-        },
-      })
-      // use a wrap statement to allow chaining onto an async function
-      cy.wrap('logging in')
-        .then(() => {
-          return new Cypress.Promise((resolve, reject) => {
-            Auth.signInWithEmailAndPassword(email, password)
-              .then((res) => resolve(res.user))
-              .catch(reject)
           })
-        })
-        // after login ensure the auth user matches expected and user menu visible
-        .its('email')
-        .should('eq', email)
-      cy.wrap(checkUI ? 'check login ui' : 'skip ui check').then(() => {
-        if (checkUI) {
-          cy.get('[data-cy=user-menu]').should('be.visible')
+        } else {
+          resolve()
         }
       })
       cy.log('user', Auth.currentUser)
@@ -199,96 +140,101 @@ const attachCustomCommands = (Cypress: Cypress.Cypress) => {
       }
     })
   })
+})
 
-  Cypress.Commands.add('deleteCurrentUser', () => {
-    return new Cypress.Promise((resolve, reject) => {
-      if (Auth.currentUser) {
-        Auth.currentUser
-          .delete()
-          .then(() => resolve(null))
-          .catch((err) => {
-            console.error(err)
-            reject(err)
-          })
-      } else {
-        resolve(null)
-      }
-    })
-  })
-
-  Cypress.Commands.add(
-    'queryDocuments',
-    (collectionName: string, fieldPath: string, opStr: any, value: string) => {
-      Cypress.log({
-        displayName: 'queryDocuments',
-        consoleProps: () => {
-          return { collectionName, fieldPath, opStr, value }
-        },
-      })
-      return firestore.queryDocuments(collectionName, fieldPath, opStr, value)
-    },
-  )
-
-  Cypress.Commands.add('step', (message: string) => {
+/**
+ * Login and logout commands use the sytem interface to log a user in or out
+ */
+Cypress.Commands.add(
+  'login',
+  (email: string, password: string, checkUI = true) => {
     Cypress.log({
-      displayName: 'step',
-      message: [`**${message}**`],
-    })
-  })
-
-  Cypress.Commands.add('toggleUserMenuOn', () => {
-    Cypress.log({ displayName: 'OPEN_USER_MENU' })
-    cy.get('[data-cy=user-menu]').should('be.exist')
-    cy.get('[data-cy=user-menu]').click()
-  })
-
-  Cypress.Commands.add('toggleUserMenuOff', () => {
-    Cypress.log({ displayName: 'CLOSE_USER_MENU' })
-    cy.get('[data-cy=header]').click({ force: true })
-  })
-
-  Cypress.Commands.add('clickMenuItem', (menuItem: UserMenuItem) => {
-    Cypress.log({
-      displayName: 'CLICK_MENU_ITEM',
+      displayName: 'login',
       consoleProps: () => {
-        return { menuItem }
+        return { email, password }
       },
     })
-    cy.toggleUserMenuOn()
-    cy.get(`[data-cy=menu-${menuItem}]`).click()
+    // use a wrap statement to allow chaining onto an async function
+    cy.wrap('logging in')
+      .then(() => {
+        return new Cypress.Promise((resolve, reject) => {
+          Auth.signInWithEmailAndPassword(email, password)
+            .then((res) => resolve(res.user))
+            .catch(reject)
+        })
+      })
+      // after login ensure the auth user matches expected and user menu visible
+      .its('email')
+      .should('eq', email)
+    cy.wrap(checkUI ? 'check login ui' : 'skip ui check').then(() => {
+      if (checkUI) {
+        cy.get('[data-cy=user-menu]').should('be.visible')
+      }
+    })
+    cy.log('user', Auth.currentUser)
+  },
+)
+
+Cypress.Commands.add('logout', (checkUI = true) => {
+  cy.wrap('logging out').then(() => {
+    return new Cypress.Promise((resolve) => {
+      Auth.signOut().then(() => resolve())
+    })
   })
-
-  Cypress.Commands.add('screenClick', () => {
-    cy.get('[data-cy=header]').click({ force: true })
+  cy.wrap(checkUI ? 'check logout ui' : 'skip ui check').then(() => {
+    if (checkUI) {
+      cy.get('[data-cy=login]')
+    }
   })
+})
 
-  Cypress.Commands.add(
-    'selectTag',
-    (tagName: string, selector = '[data-cy=tag-select]') => {
-      cy.log('select tag', tagName)
-      cy.get(`${selector} input`)
-        .click({ force: true })
-        .type(tagName, { force: true })
-        .get(`${selector} .data-cy__menu-list`)
-        .contains(tagName)
-        .click()
-    },
-  )
-
-  Cypress.Commands.add('interceptAddressSearchFetch', (addressResponse) => {
-    cy.intercept('GET', 'https://nominatim.openstreetmap.org/search*', {
-      body: addressResponse,
-    }).as('fetchAddress')
+Cypress.Commands.add('deleteCurrentUser', () => {
+  return new Cypress.Promise((resolve, reject) => {
+    if (Auth.currentUser) {
+      Auth.currentUser
+        .delete()
+        .then(() => resolve(null))
+        .catch((err) => {
+          console.error(err)
+          reject(err)
+        })
+    } else {
+      resolve(null)
+    }
   })
+})
 
-  Cypress.Commands.add('interceptAddressReverseFetch', (addressResponse) => {
-    cy.intercept('GET', 'https://nominatim.openstreetmap.org/reverse?*', {
-      body: addressResponse,
-    }).as('fetchAddressReverse')
+Cypress.Commands.add(
+  'queryDocuments',
+  (collectionName: string, fieldPath: string, opStr: any, value: string) => {
+    Cypress.log({
+      displayName: 'queryDocuments',
+      consoleProps: () => {
+        return { collectionName, fieldPath, opStr, value }
+      },
+    })
+    return firestore.queryDocuments(collectionName, fieldPath, opStr, value)
+  },
+)
+
+Cypress.Commands.add('step', (message: string) => {
+  Cypress.log({
+    displayName: 'step',
+    message: [`**${message}**`],
   })
-}
+})
 
-attachCustomCommands(Cypress)
+Cypress.Commands.add('interceptAddressSearchFetch', (addressResponse) => {
+  cy.intercept('GET', 'https://nominatim.openstreetmap.org/search*', {
+    body: addressResponse,
+  }).as('fetchAddress')
+})
+
+Cypress.Commands.add('interceptAddressReverseFetch', (addressResponse) => {
+  cy.intercept('GET', 'https://nominatim.openstreetmap.org/reverse?*', {
+    body: addressResponse,
+  }).as('fetchAddressReverse')
+})
 
 /**
  * Overwrite default logging to also output to console
