+++ packages/cypress/src/support/index.ts
@@ -14,5 +14,6 @@
 // ***********************************************************
 import './hooks'
 import './commands'
+import './commandsUi'
 import './CustomAssertations'
 import './rules'
